third_party/WebKit/LayoutTests/http/tests/navigation/hash-navigation-in-unload-handler.html

<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
var test = async_test("Tests that a hash navigation in the unload handler will not block the initial navigation");
window.onload = function() {
    var i = document.createElement('iframe');
    document.body.appendChild(i);
    i.contentWindow.onunload = function () {
        i.contentWindow.location.hash = '#hash';
    }
    setTimeout(function() {
        i.contentWindow.location.assign('about:blank');
        setTimeout(function() {
            test.step(function() { assert_true(!i.contentWindow.location.hash && i.contentWindow.location.href == 'about:blank'); });
            test.done();
        }, 100);
    }, 100);
}
</script>
</head>
<body>
</body>
</html>
